In case graceful shutdown failed, fallback to stopping VM immediately
Checking capabilities fails if guest has crashed or failed to boot in the first
place. In that case vagrant destroy fails to do it's job:

==> default: Halting VM...
ERROR warden: Error occurred: Guest-specific operations were attempted on a machine that is not
ready for guest communication. This should not happen and a bug
should be reported.