Add "vmname" provider option
that represents the name that oVirt should attach to the VM, as distinct
from its hostname.

Inspired by the hyperv provider's "vmname" option; see:

    https://github.com/hashicorp/vagrant/blob/3d68e16f1b5d616f879d24f48dac2598acac0a38/plugins/providers/hyperv/config.rb#L28-L29

If "vmname" is unspecified, use "config.vm.hostname"; if this is
unspecified, use the default value "vagrant".

This changeset introduces the utility module "Util::MachineNames" that
provides the instance method "#machine_vmname" for inferring the vmname
of a Vagrant-managed machine, and the instance method
"#machine_hostname" for inferring the hostname.

Motivation: under the hood, libvirt-based machines have to abide by
systemd-machined's machine naming rules, including a 64-character length
limit. systemd-machined also requires machine names to be valid
hostnames, which means (among other things) that they cannot end in a
".".  Older versions of libvirt do not check for a trailing "." when
truncating machine names to 64 characters, which leads to a fatal error
creating a machine when the 64th character happens to be ".".  The
"vmname" setting helps because we can set it to a value that contains no
dots, while still setting the machine *hostname* to a value that does
contain dots.